# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.AdminLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.AdminLoginResponseDoc:
    properties:
      admin:
        $ref: '#/definitions/models.Admin'
      session:
        $ref: '#/definitions/handler.TokenResponseDoc'
    type: object
...
  handler.CreateAdminRequest:
    properties:
      email:
        type: string
      password:
        type: string
      property_id:
        description: bisa kosong kalau mau super admin
        type: string
      role:
        type: string
    type: object
...
  handler.TokenResponseDoc:
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        additionalProperties: true
        type: object
    type: object
  handler.UpdateAdminRequest:
    properties:
      is_active:
        type: boolean
      property_id:
        type: string
      role:
        type: string
    type: object
  handler.UpdateBookingStatusRequest:
    properties:
      note:
        type: string
      refund_amount:
        type: number
      status:
        $ref: '#/definitions/models.BookingStatus'
    type: object
...
  models.Admin:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      property_id:
        type: string
      role:
        type: string
    type: object
  models.Booking:
    properties:
      booking_status:
        $ref: '#/definitions/models.BookingStatus'
      check_in:
        type: string
      check_out:
        type: string
      created_at:
        type: string
      guest_id:
        type: string
      id:
        type: string
      nights:
        type: integer
      note:
        type: string
      property_id:
        type: string
      refund_amount:
        type: number
      room_id:
        type: string
      total_price:
        type: number
    type: object
  models.BookingStatus:
    enum:
    - New
    - Confirmed
    - Cancelled
    - CheckedIn
    - CheckedOut
    - NoShow
    type: string
    x-enum-varnames:
    - BookingStatusNew
    - BookingStatusConfirmed
    - BookingStatusCancel
    - BookingStatusCheckedIn
    - BookingStatusCheckedOut
    - BookingStatusNoShow
...
info:
  contact: {}
  description: REST API for hotel booking management (guest, booking, admin inventory,
    reports).
  title: Hotel Booking API
  version: "1.0"
paths:
  /admin/bookings:
    get:
      parameters:
      - description: Property ID
        in: query
        name: property_id
        type: string
      - description: Booking status
        in: query
        name: status
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Booking'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List bookings
      tags:
      - Bookings
  /admin/bookings/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Update booking status
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateBookingStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update booking status
      tags:
      - Bookings
...
  /admin/users:
    get:
      parameters:
      - description: Property ID
        in: query
        name: property_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Admin'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List admin users
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Create admin
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Admin'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create admin user
      tags:
      - Admin
  /admin/users/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      - description: Update admin
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update admin user
      tags:
      - Admin
  /admin/users/{id}/activate:
    post:
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Activate admin user
      tags:
      - Admin
  /admin/users/{id}/deactivate:
    post:
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate admin user
      tags:
      - Admin
  /auth/admin/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Admin login
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AdminLoginResponseDoc'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login admin
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"